requirements.txt
torch>=1.8.0
torch-geometric>=2.0.0
numpy>=1.19.0
pandas>=1.1.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
networkx>=2.5.0
tqdm>=4.50.0
pyyaml>=5.4.0
transformers>=4.5.0
tensorboard>=2.4.0
scipy>=1.6.0
